+++ src/app/search/search.module.scss
@@ -164,100 +164,3 @@ div.notFound {
     }
   }
 }
-
-.btnWrap {
-  display: flex;
-  gap: 0.8rem;
-  margin: 10rem 0;
-
-  .btnPrev,
-  .btnNext {
-    width: 5rem;
-    height: 5rem;
-    padding: 0;
-    border: none;
-    background-color: var(--background);
-
-    span {
-      display: none;
-    }
-
-    &:not(:disabled):active {
-      background-color: var(--primary);
-      border-color: var(--primary);
-      svg path {
-        fill: #fff;
-      }
-    }
-
-    @media (hover: hover) {
-      &:not(:disabled):hover {
-        display: flex;
-        align-items: center;
-        width: auto;
-        min-width: 5rem;
-        padding: 0 1.2rem;
-        background-color: var(--primary);
-        border-color: var(--primary);
-        color: #fff;
-
-        svg path {
-          fill: #fff;
-        }
-
-        span {
-          display: flex;
-        }
-      }
-    }
-  }
-  .btnPrev span {
-    margin-right: 1rem;
-  }
-
-  .btnNext span {
-    margin-left: 1rem;
-  }
-
-  .pageNav {
-    display: flex;
-  }
-
-  .pageNav {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    label {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      width: 4.2rem;
-      height: 4.2rem;
-      color: var(--grayLv3);
-      border-bottom: 4px solid transparent;
-      cursor: pointer;
-    }
-
-    input:checked + label {
-      color: var(--primary);
-      background-color: var(--activation);
-      border-bottom: 4px solid var(--primary);
-      box-sizing: border-box;
-    }
-
-    .ellipsis {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-
-      p {
-        display: flex;
-        justify-content: center;
-        width: 4.2rem;
-        font-size: 1.6rem;
-        color: var(--grayLv3);
-      }
-    }
-  }
-}
